+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Added utils method to create message. Added example profile to add some messages.
+  [sgeulette]
 
 
 0.1 (2015-09-08)